Refactored the key binding system and implemented generic procedures (#237)
* Refactored the key binding system and implemented generic procedures
 - vim bindings
 - alternative bindings
 - register cmds before extension cmds
 - add subcommand invocation by name

I took the emacs approach where every functionality is a certain defined
procedure which can, but doesn't have to, be bound to a key. Each can
also be given a name so the future implementation of subcommands should
be very easy. I chose the name 'procedure' as there was already a lot
of ambiguity with the word 'command'.

I kept all the existing key bindings, although I modified some to align
them with vim. I also added a incomplete emacs profile just to showcase
how bindings can be changed.