sheet: Override header span and skip rows when using properties file (#266)
* sheet: Override header span and skip rows when using properties file

The properties file was overriding opts.rows_to_ignore and
opts.header_span values that need to be set to zero when using an
index. This recreates the parser with zsv_new after the opts have
been rewritten when an index is active.

* sheet/index: Fix building the index with a filter

Previously the indexer was using the file offsets from the original
file when the filter was active. Instead it should use offsets in the
temporary filtered file.

The writer library doesn't record or expose the total number of bytes
written and flushing to the stream and retrieving the offset would be
expensive and invasive. So this splits filtering and indexing into two
operations.

* Update ci.yml

add CI tests for macos-14, which should pass

---------

Co-authored-by: liquidaty <info@liquidaty.com>